Revisited naming of files, classes and interfaces.
