+++ src/enjug/erijan/games/yatzy/RollControl.java
@@ -3,7 +3,7 @@ package enjug.erijan.games.yatzy;
 /**
  * Created by Jan Eriksson on 16/11/15.
  */
-public interface RollingControlling {
+public interface RollControl {
   void rollActiveDice();
   int rollsLeft();
 }
