+++ src/enjug/erijan/games/yatzy/rules/YahtzeeBoxes.java
@@ -1,6 +1,6 @@
 package enjug.erijan.games.yatzy.rules;
 
-import enjug.erijan.games.yatzy.YatzyVariants;
+import enjug.erijan.games.yatzy.VariantsFactoryImpl;
 
 /**
  * Created by Jan Eriksson on 27/10/15.
@@ -25,7 +25,7 @@ public enum YahtzeeBoxes implements ScoreRule, ScoreBoxFactory {
   YATZY (result -> ScoreCalculator.nSame(5, 6, result)),
   TOTAL (result -> ScoreCalculator.totalSum(result));
 
-  private static final YatzyVariants yatzyVariant = YatzyVariants.YATZY;
+  private static final VariantsFactoryImpl yatzyVariant = VariantsFactoryImpl.YATZY;
 
   private final ScoreRule scoreRule;
 
@@ -33,7 +33,7 @@ public enum YahtzeeBoxes implements ScoreRule, ScoreBoxFactory {
     this.scoreRule = scoreRule;
   }
 
-  public static YatzyVariants getYatzyVariant() {
+  public static VariantsFactoryImpl getYatzyVariant() {
     return yatzyVariant;
   }
